+++ src/app/Http/Controllers/ReservationController.php
@@ -153,6 +153,7 @@ class ReservationController extends Controller
         $reservation = Reservation::create($request->all());
         $reservation = $reservation->load(['guest', 'room']);
         $reservation->created_by = auth()->id();
+        $reservation->save();
 
         return redirect()->route('reservations.index')->with([
             'status' => 'success',
@@ -230,6 +231,7 @@ class ReservationController extends Controller
         try {
             $reservation->update($request->only(['room_id', 'guest_id', 'daily_price', 'scheduled_check_in', 'scheduled_check_out']));
             $reservation->updated_by = auth()->id();
+            $reservation->save();
 
             return redirect()->route('reservations.update', $id)->with([
                 'status' => 'success',
@@ -251,6 +253,15 @@ class ReservationController extends Controller
      */
     public function destroy(Reservation $reservation)
     {
+        // No nível de operador não é mais possível apagar reservas após o check-in
+        if (!auth()->user()->can('delete', $reservation)) {
+            return redirect()->route('reservations.show', $id)->with([
+                'status' => 'error',
+                'alert-type' => 'danger',
+                'message' => "Você não pode mais apagar esta reserva.",
+            ]);
+        }
+
         $reservation->delete();
 
         return redirect()->back()->with([
